# backend/domain/models/contract.py
from datetime import datetime
from typing import Literal
from uuid import UUID

from pydantic import BaseModel, ConfigDict


class Contract(BaseModel):
    model_config = ConfigDict(from_attributes=True)
    id: UUID
    deal_id: UUID | None = None
    creator_id: UUID | None = None
    file_path: str | None = None
    status: str
    usage_rights_duration_days: int | None = None
    exclusivity_scope: str | None = None
    revision_limit: int | None = None
    payment_timeline_days: int | None = None
    previous_status: str | None = None
    raw_text: str | None = None
    created_at: datetime
    updated_at: datetime

class ContractStatusUpdateRequest(BaseModel):
    status: Literal['uploaded', 'signed', 'void']