# openapi.json
{"openapi": "3.1.0", "info": {"title": "Creator OS", "version": "0.1.0"}, "paths": {"/creators": {"post": {"tags": ["creators"], "summary": "Onboard Creator", "operationId": "onboard_creator_creators_post", "parameters": [{"name": "agency_id", "in": "query", "required": true, "schema": {"type": "string", "format": "uuid", "title": "Agency Id"}}, {"name": "actor_id", "in": "query", "required": true, "schema": {"type": "string", "format": "uuid", "title": "Actor Id"}}, {"name": "actor_type", "in": "query", "required": true, "schema": {"type": "string", "title": "Actor Type"}}], "requestBody": {"required": true, "content": {"application/json": {"schema": {"$ref": "#/components/schemas/CreatorOnboardRequest"}}}}, "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/Creator"}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}}, "/creators/{id}": {"get": {"tags": ["creators"], "summary": "Get Creator", "operationId": "get_creator_creators__id__get", "parameters": [{"name": "id", "in": "path", "required": true, "schema": {"type": "string", "format": "uuid", "title": "Id"}}], "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/Creator"}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}}, "/deals/{deal_id}/contracts": {"post": {"tags": ["contracts"], "summary": "Upload Contract", "operationId": "upload_contract_deals__deal_id__contracts_post", "parameters": [{"name": "deal_id", "in": "path", "required": true, "schema": {"type": "string", "format": "uuid", "title": "Deal Id"}}], "requestBody": {"required": true, "content": {"multipart/form-data": {"schema": {"$ref": "#/components/schemas/Body_upload_contract_deals__deal_id__contracts_post"}}}}, "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}}, "/contracts/{id}/status": {"patch": {"tags": ["contracts"], "summary": "Update Contract Status", "operationId": "update_contract_status_contracts__id__status_patch", "parameters": [{"name": "id", "in": "path", "required": true, "schema": {"type": "string", "format": "uuid", "title": "Id"}}, {"name": "agency_id", "in": "query", "required": true, "schema": {"type": "string", "format": "uuid", "title": "Agency Id"}}, {"name": "actor_id", "in": "query", "required": true, "schema": {"type": "string", "format": "uuid", "title": "Actor Id"}}, {"name": "actor_type", "in": "query", "required": true, "schema": {"type": "string", "title": "Actor Type"}}], "requestBody": {"required": true, "content": {"application/json": {"schema": {"$ref": "#/components/schemas/ContractStatusUpdateRequest"}}}}, "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/Contract"}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}}, "/deals/{deal_id}/deliverables": {"post": {"tags": ["deliverables"], "summary": "Create Deliverable", "operationId": "create_deliverable_deals__deal_id__deliverables_post", "parameters": [{"name": "deal_id", "in": "path", "required": true, "schema": {"type": "string", "format": "uuid", "title": "Deal Id"}}, {"name": "agency_id", "in": "query", "required": true, "schema": {"type": "string", "format": "uuid", "title": "Agency Id"}}, {"name": "actor_id", "in": "query", "required": true, "schema": {"type": "string", "format": "uuid", "title": "Actor Id"}}, {"name": "actor_type", "in": "query", "required": true, "schema": {"type": "string", "title": "Actor Type"}}], "requestBody": {"required": true, "content": {"application/json": {"schema": {"$ref": "#/components/schemas/DeliverableCreateRequest"}}}}, "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/Deliverable"}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}}, "/deliverables/{id}/status": {"patch": {"tags": ["deliverables"], "summary": "Update Deliverable Status", "operationId": "update_deliverable_status_deliverables__id__status_patch", "parameters": [{"name": "id", "in": "path", "required": true, "schema": {"type": "string", "format": "uuid", "title": "Id"}}, {"name": "deal_id", "in": "query", "required": true, "schema": {"type": "string", "format": "uuid", "title": "Deal Id"}}, {"name": "agency_id", "in": "query", "required": true, "schema": {"type": "string", "format": "uuid", "title": "Agency Id"}}, {"name": "actor_id", "in": "query", "required": true, "schema": {"type": "string", "format": "uuid", "title": "Actor Id"}}, {"name": "actor_type", "in": "query", "required": true, "schema": {"type": "string", "title": "Actor Type"}}], "requestBody": {"required": true, "content": {"application/json": {"schema": {"$ref": "#/components/schemas/DeliverableStatusUpdateRequest"}}}}, "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/Deliverable"}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}}, "/activity-log": {"get": {"tags": ["activity-log"], "summary": "Get Activity Logs", "operationId": "get_activity_logs_activity_log_get", "parameters": [{"name": "deal_id", "in": "query", "required": true, "schema": {"type": "string", "format": "uuid", "title": "Deal Id"}}], "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {"type": "array", "items": {"$ref": "#/components/schemas/ActivityLog"}, "title": "Response Get Activity Logs Activity Log Get"}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}}, "/negotiation/sessions": {"post": {"tags": ["Negotiation"], "summary": "Create Session", "operationId": "create_session_negotiation_sessions_post", "parameters": [{"name": "agency_id", "in": "query", "required": true, "schema": {"type": "string", "format": "uuid", "title": "Agency Id"}}], "requestBody": {"required": true, "content": {"application/json": {"schema": {"$ref": "#/components/schemas/NegotiationSessionCreate"}}}}, "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/NegotiationSession"}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}}, "/negotiation/sessions/{id}/rate-range": {"get": {"tags": ["Negotiation"], "summary": "Get Rate Range", "operationId": "get_rate_range_negotiation_sessions__id__rate_range_get", "parameters": [{"name": "id", "in": "path", "required": true, "schema": {"type": "string", "format": "uuid", "title": "Id"}}], "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/RateRangeOut"}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}}, "/negotiation/sessions/{id}/forecast": {"get": {"tags": ["Negotiation"], "summary": "Get Forecast", "operationId": "get_forecast_negotiation_sessions__id__forecast_get", "parameters": [{"name": "id", "in": "path", "required": true, "schema": {"type": "string", "format": "uuid", "title": "Id"}}, {"name": "weekly_growth_rate", "in": "query", "required": false, "schema": {"type": "number", "default": 0.05, "title": "Weekly Growth Rate"}}, {"name": "weeks", "in": "query", "required": false, "schema": {"type": "integer", "default": 12, "title": "Weeks"}}], "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {"type": "array", "items": {"$ref": "#/components/schemas/ForecastPoint"}, "title": "Response Get Forecast Negotiation Sessions  Id  Forecast Get"}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}}, "/negotiation/sessions/{id}/checklist": {"patch": {"tags": ["Negotiation"], "summary": "Update Checklist", "operationId": "update_checklist_negotiation_sessions__id__checklist_patch", "parameters": [{"name": "id", "in": "path", "required": true, "schema": {"type": "string", "format": "uuid", "title": "Id"}}], "requestBody": {"required": true, "content": {"application/json": {"schema": {"$ref": "#/components/schemas/ChecklistState"}}}}, "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/NegotiationSession"}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}}, "/negotiation/sessions/{id}/counter-offer": {"post": {"tags": ["Negotiation"], "summary": "Create Counter Offer", "operationId": "create_counter_offer_negotiation_sessions__id__counter_offer_post", "parameters": [{"name": "id", "in": "path", "required": true, "schema": {"type": "string", "format": "uuid", "title": "Id"}}, {"name": "actor_id", "in": "query", "required": true, "schema": {"type": "string", "format": "uuid", "title": "Actor Id"}}, {"name": "actor_type", "in": "query", "required": true, "schema": {"type": "string", "title": "Actor Type"}}, {"name": "agency_id", "in": "query", "required": true, "schema": {"type": "string", "format": "uuid", "title": "Agency Id"}}], "requestBody": {"required": true, "content": {"application/json": {"schema": {"$ref": "#/components/schemas/CounterOfferRequest"}}}}, "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/NegotiationOffer"}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}}, "/payments/{contract_id}/mark-paid": {"patch": {"tags": ["Payments"], "summary": "Mark Paid", "operationId": "mark_paid_payments__contract_id__mark_paid_patch", "parameters": [{"name": "contract_id", "in": "path", "required": true, "schema": {"type": "string", "format": "uuid", "title": "Contract Id"}}, {"name": "payment_id", "in": "query", "required": true, "schema": {"type": "string", "format": "uuid", "title": "Payment Id"}}, {"name": "agency_id", "in": "query", "required": true, "schema": {"type": "string", "format": "uuid", "title": "Agency Id"}}, {"name": "actor_id", "in": "query", "required": true, "schema": {"type": "string", "format": "uuid", "title": "Actor Id"}}, {"name": "actor_type", "in": "query", "required": true, "schema": {"type": "string", "title": "Actor Type"}}, {"name": "deal_id", "in": "query", "required": true, "schema": {"type": "string", "format": "uuid", "title": "Deal Id"}}], "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/Payment"}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}}, "/deals/{deal_id}/brand-portal-link": {"post": {"tags": ["brand_portal"], "summary": "Create Brand Portal Link", "description": "Creates a new brand portal link for a deal.\nThis endpoint requires agency authentication in reality, but here it's just a router.", "operationId": "create_brand_portal_link_deals__deal_id__brand_portal_link_post", "parameters": [{"name": "deal_id", "in": "path", "required": true, "schema": {"type": "string", "format": "uuid", "title": "Deal Id"}}], "requestBody": {"required": true, "content": {"application/json": {"schema": {"$ref": "#/components/schemas/BrandPortalTokenCreate"}}}}, "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {"type": "object", "additionalProperties": true, "title": "Response Create Brand Portal Link Deals  Deal Id  Brand Portal Link Post"}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}}, "/portal/{token}/summary": {"get": {"tags": ["brand_portal"], "summary": "Get Portal Summary", "description": "Gets the summary for a brand portal, token gated.", "operationId": "get_portal_summary_portal__token__summary_get", "parameters": [{"name": "token", "in": "path", "required": true, "schema": {"type": "string", "title": "Token"}}], "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/BrandPortalSummaryOut"}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}}, "/portal/{token}/approve": {"post": {"tags": ["brand_portal"], "summary": "Approve Deliverable", "description": "Approves or rejects a deliverable via the brand portal.", "operationId": "approve_deliverable_portal__token__approve_post", "parameters": [{"name": "token", "in": "path", "required": true, "schema": {"type": "string", "title": "Token"}}], "requestBody": {"required": true, "content": {"application/json": {"schema": {"$ref": "#/components/schemas/BrandApprovalRequest"}}}}, "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}}}, "components": {"schemas": {"ActivityLog": {"properties": {"id": {"type": "string", "format": "uuid", "title": "Id"}, "agency_id": {"type": "string", "format": "uuid", "title": "Agency Id"}, "deal_id": {"anyOf": [{"type": "string", "format": "uuid"}, {"type": "null"}], "title": "Deal Id"}, "actor_type": {"type": "string", "title": "Actor Type"}, "actor_label": {"type": "string", "title": "Actor Label"}, "action": {"type": "string", "title": "Action"}, "metadata": {"additionalProperties": true, "type": "object", "title": "Metadata", "default": {}}, "created_at": {"type": "string", "format": "date-time", "title": "Created At"}}, "type": "object", "required": ["id", "agency_id", "actor_type", "actor_label", "action", "created_at"], "title": "ActivityLog"}, "Body_upload_contract_deals__deal_id__contracts_post": {"properties": {"file": {"type": "string", "contentMediaType": "application/octet-stream", "title": "File"}}, "type": "object", "required": ["file"], "title": "Body_upload_contract_deals__deal_id__contracts_post"}, "BrandApprovalRequest": {"properties": {"deliverable_id": {"type": "string", "format": "uuid", "title": "Deliverable Id"}, "decision": {"type": "string", "enum": ["approved", "revision_requested"], "title": "Decision"}, "note": {"anyOf": [{"type": "string"}, {"type": "null"}], "title": "Note"}}, "type": "object", "required": ["deliverable_id", "decision"], "title": "BrandApprovalRequest"}, "BrandPortalSummaryOut": {"properties": {"deal": {"$ref": "#/components/schemas/Deal"}, "deliverables": {"items": {"$ref": "#/components/schemas/Deliverable"}, "type": "array", "title": "Deliverables"}, "activity_log": {"items": {"$ref": "#/components/schemas/ActivityLog"}, "type": "array", "title": "Activity Log"}}, "type": "object", "required": ["deal", "deliverables", "activity_log"], "title": "BrandPortalSummaryOut"}, "BrandPortalTokenCreate": {"properties": {"deal_id": {"type": "string", "format": "uuid", "title": "Deal Id"}, "expires_in_days": {"type": "integer", "title": "Expires In Days"}}, "type": "object", "required": ["deal_id", "expires_in_days"], "title": "BrandPortalTokenCreate"}, "ChecklistState": {"properties": {"usage_rights_duration_set": {"type": "boolean", "title": "Usage Rights Duration Set", "default": false}, "exclusivity_scope_set": {"type": "boolean", "title": "Exclusivity Scope Set", "default": false}, "revision_limit_set": {"type": "boolean", "title": "Revision Limit Set", "default": false}, "payment_timeline_set": {"type": "boolean", "title": "Payment Timeline Set", "default": false}}, "type": "object", "title": "ChecklistState"}, "Contract": {"properties": {"id": {"type": "string", "format": "uuid", "title": "Id"}, "deal_id": {"type": "string", "format": "uuid", "title": "Deal Id"}, "file_path": {"anyOf": [{"type": "string"}, {"type": "null"}], "title": "File Path"}, "status": {"type": "string", "title": "Status"}, "usage_rights_duration_days": {"anyOf": [{"type": "integer"}, {"type": "null"}], "title": "Usage Rights Duration Days"}, "exclusivity_scope": {"anyOf": [{"type": "string"}, {"type": "null"}], "title": "Exclusivity Scope"}, "revision_limit": {"anyOf": [{"type": "integer"}, {"type": "null"}], "title": "Revision Limit"}, "payment_timeline_days": {"anyOf": [{"type": "integer"}, {"type": "null"}], "title": "Payment Timeline Days"}, "created_at": {"type": "string", "format": "date-time", "title": "Created At"}, "updated_at": {"type": "string", "format": "date-time", "title": "Updated At"}}, "type": "object", "required": ["id", "deal_id", "status", "created_at", "updated_at"], "title": "Contract"}, "ContractStatusUpdateRequest": {"properties": {"status": {"type": "string", "enum": ["uploaded", "signed", "void"], "title": "Status"}}, "type": "object", "required": ["status"], "title": "ContractStatusUpdateRequest"}, "CounterOfferRequest": {"properties": {"amount": {"type": "number", "title": "Amount"}, "message": {"anyOf": [{"type": "string"}, {"type": "null"}], "title": "Message"}}, "type": "object", "required": ["amount"], "title": "CounterOfferRequest"}, "Creator": {"properties": {"id": {"type": "string", "format": "uuid", "title": "Id"}, "agency_id": {"type": "string", "format": "uuid", "title": "Agency Id"}, "user_id": {"anyOf": [{"type": "string", "format": "uuid"}, {"type": "null"}], "title": "User Id"}, "display_name": {"type": "string", "title": "Display Name"}, "instagram_handle": {"type": "string", "title": "Instagram Handle"}, "niche": {"type": "string", "title": "Niche"}, "follower_tier": {"type": "string", "title": "Follower Tier"}, "created_at": {"type": "string", "format": "date-time", "title": "Created At"}}, "type": "object", "required": ["id", "agency_id", "display_name", "instagram_handle", "niche", "follower_tier", "created_at"], "title": "Creator"}, "CreatorOnboardRequest": {"properties": {"display_name": {"type": "string", "title": "Display Name"}, "instagram_handle": {"type": "string", "title": "Instagram Handle"}, "niche": {"type": "string", "title": "Niche"}, "follower_tier": {"type": "string", "title": "Follower Tier"}}, "type": "object", "required": ["display_name", "instagram_handle", "niche", "follower_tier"], "title": "CreatorOnboardRequest"}, "Deal": {"properties": {"id": {"type": "string", "format": "uuid", "title": "Id"}, "agency_id": {"type": "string", "format": "uuid", "title": "Agency Id"}, "creator_id": {"type": "string", "format": "uuid", "title": "Creator Id"}, "brand_name": {"type": "string", "title": "Brand Name"}, "brand_contact_email": {"type": "string", "title": "Brand Contact Email"}, "status": {"type": "string", "title": "Status"}, "created_at": {"type": "string", "format": "date-time", "title": "Created At"}, "updated_at": {"type": "string", "format": "date-time", "title": "Updated At"}}, "type": "object", "required": ["id", "agency_id", "creator_id", "brand_name", "brand_contact_email", "status", "created_at", "updated_at"], "title": "Deal"}, "Deliverable": {"properties": {"id": {"type": "string", "format": "uuid", "title": "Id"}, "deal_id": {"type": "string", "format": "uuid", "title": "Deal Id"}, "title": {"type": "string", "title": "Title"}, "description": {"type": "string", "title": "Description"}, "status": {"type": "string", "title": "Status"}, "file_path": {"anyOf": [{"type": "string"}, {"type": "null"}], "title": "File Path"}, "created_at": {"type": "string", "format": "date-time", "title": "Created At"}, "updated_at": {"type": "string", "format": "date-time", "title": "Updated At"}}, "type": "object", "required": ["id", "deal_id", "title", "description", "status", "created_at", "updated_at"], "title": "Deliverable"}, "DeliverableCreateRequest": {"properties": {"deal_id": {"type": "string", "format": "uuid", "title": "Deal Id"}, "title": {"type": "string", "title": "Title"}, "description": {"type": "string", "title": "Description"}}, "type": "object", "required": ["deal_id", "title", "description"], "title": "DeliverableCreateRequest"}, "DeliverableStatusUpdateRequest": {"properties": {"status": {"type": "string", "enum": ["submitted", "approved", "rejected", "revision_requested"], "title": "Status"}, "note": {"anyOf": [{"type": "string"}, {"type": "null"}], "title": "Note"}}, "type": "object", "required": ["status"], "title": "DeliverableStatusUpdateRequest"}, "ForecastPoint": {"properties": {"week": {"type": "integer", "title": "Week"}, "low": {"type": "number", "title": "Low"}, "mid": {"type": "number", "title": "Mid"}, "high": {"type": "number", "title": "High"}}, "type": "object", "required": ["week", "low", "mid", "high"], "title": "ForecastPoint"}, "HTTPValidationError": {"properties": {"detail": {"items": {"$ref": "#/components/schemas/ValidationError"}, "type": "array", "title": "Detail"}}, "type": "object", "title": "HTTPValidationError"}, "NegotiationOffer": {"properties": {"id": {"type": "string", "format": "uuid", "title": "Id"}, "session_id": {"type": "string", "format": "uuid", "title": "Session Id"}, "actor_type": {"type": "string", "title": "Actor Type"}, "actor_id": {"type": "string", "format": "uuid", "title": "Actor Id"}, "amount": {"type": "number", "title": "Amount"}, "message": {"anyOf": [{"type": "string"}, {"type": "null"}], "title": "Message"}, "created_at": {"type": "string", "format": "date-time", "title": "Created At"}}, "type": "object", "required": ["id", "session_id", "actor_type", "actor_id", "amount", "created_at"], "title": "NegotiationOffer"}, "NegotiationSession": {"properties": {"id": {"type": "string", "format": "uuid", "title": "Id"}, "deal_id": {"type": "string", "format": "uuid", "title": "Deal Id"}, "status": {"type": "string", "title": "Status"}, "base_rate": {"type": "number", "title": "Base Rate"}, "range_low": {"type": "number", "title": "Range Low"}, "range_high": {"type": "number", "title": "Range High"}, "checklist": {"$ref": "#/components/schemas/ChecklistState"}, "created_at": {"type": "string", "format": "date-time", "title": "Created At"}, "updated_at": {"type": "string", "format": "date-time", "title": "Updated At"}}, "type": "object", "required": ["id", "deal_id", "status", "base_rate", "range_low", "range_high", "checklist", "created_at", "updated_at"], "title": "NegotiationSession"}, "NegotiationSessionCreate": {"properties": {"deal_id": {"type": "string", "format": "uuid", "title": "Deal Id"}, "views_per_week": {"type": "number", "title": "Views Per Week"}, "niche_cpm": {"type": "number", "title": "Niche Cpm"}, "follower_tier_multiplier": {"type": "number", "title": "Follower Tier Multiplier"}, "engagement_rate_adjustment": {"type": "number", "title": "Engagement Rate Adjustment"}}, "type": "object", "required": ["deal_id", "views_per_week", "niche_cpm", "follower_tier_multiplier", "engagement_rate_adjustment"], "title": "NegotiationSessionCreate"}, "Payment": {"properties": {"id": {"type": "string", "format": "uuid", "title": "Id"}, "contract_id": {"type": "string", "format": "uuid", "title": "Contract Id"}, "amount_inr": {"type": "string", "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$", "title": "Amount Inr"}, "status": {"type": "string", "title": "Status"}, "invoice_number": {"type": "string", "title": "Invoice Number"}, "invoice_pdf_path": {"anyOf": [{"type": "string"}, {"type": "null"}], "title": "Invoice Pdf Path"}, "due_date": {"anyOf": [{"type": "string", "format": "date"}, {"type": "null"}], "title": "Due Date"}, "paid_at": {"anyOf": [{"type": "string", "format": "date-time"}, {"type": "null"}], "title": "Paid At"}, "created_at": {"type": "string", "format": "date-time", "title": "Created At"}, "updated_at": {"type": "string", "format": "date-time", "title": "Updated At"}}, "type": "object", "required": ["id", "contract_id", "amount_inr", "status", "invoice_number", "created_at", "updated_at"], "title": "Payment"}, "RateRangeOut": {"properties": {"base_rate": {"type": "number", "title": "Base Rate"}, "range_low": {"type": "number", "title": "Range Low"}, "range_high": {"type": "number", "title": "Range High"}}, "type": "object", "required": ["base_rate", "range_low", "range_high"], "title": "RateRangeOut"}, "ValidationError": {"properties": {"loc": {"items": {"anyOf": [{"type": "string"}, {"type": "integer"}]}, "type": "array", "title": "Location"}, "msg": {"type": "string", "title": "Message"}, "type": {"type": "string", "title": "Error Type"}, "input": {"title": "Input"}, "ctx": {"type": "object", "title": "Context"}}, "type": "object", "required": ["loc", "msg", "type"], "title": "ValidationError"}}}}